+++ tests/files/src/0582_declared_nullable_return.php
@@ -1,0 +1,18 @@
+<?php
+
+class Foo582 {
+    /**
+     * Should emit PhanTypeMismatchDeclaredReturn
+     * @return ?stdClass some description
+     */
+    public function getNullable() : stdClass {
+        return new stdClass();
+    }
+    /**
+     * Should emit PhanTypeMismatchDeclaredReturn
+     * @return stdClass|false some description
+     */
+    public function getFalseable() : stdClass {
+        return new stdClass();
+    }
+}